// frontend/src/app/core/api/generated-types/Match.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Mecz w ramach sezonu i rozgrywek.
 */
export type Match = { id: string, seasonId: string, competitionId: string, opponent: string, 
/**
 * Data meczu `RRRR-MM-DD`.
 */
date: string, 
/**
 * Mecz u siebie (`true`) czy na wyjeździe.
 */
home: boolean, };